move wrapJsxElementAsComponent
